+++ Assets/Scripts/UI/Inventory/EquipmentDetailWindow.cs
@@ -107,7 +107,7 @@ public class EquipmentDetailWindow : MonoBehaviour
         string physDPS;
         string physDamage = "";
         MinMaxRange range;
-        double dps;
+        float dps;
 
         for (int i = 1; i < (int)ElementType.DIVINE; i++)
         {
@@ -115,7 +115,7 @@ public class EquipmentDetailWindow : MonoBehaviour
             if (!range.IsZero())
             {
                 hasElemental = true;
-                dps = (range.min + range.max) / 2d * weaponItem.AttackSpeed;
+                dps = (range.min + range.max) / 2f * weaponItem.AttackSpeed;
                 elementalDps.Add(LocalizationManager.BuildElementalDamageString(dps.ToString("F2"), (ElementType)i));
                 elementalDamage.Add(LocalizationManager.BuildElementalDamageString(range.min + "-" + range.max, (ElementType)i));
             }
@@ -125,7 +125,7 @@ public class EquipmentDetailWindow : MonoBehaviour
         if (!range.IsZero())
         {
             hasPrimordial = true;
-            dps = (range.min + range.max) / 2d * weaponItem.AttackSpeed;
+            dps = (range.min + range.max) / 2f * weaponItem.AttackSpeed;
             primDps.Add(LocalizationManager.BuildElementalDamageString(dps.ToString("F2"), ElementType.DIVINE));
             primDamage.Add(LocalizationManager.BuildElementalDamageString(range.min + "-" + range.max, ElementType.DIVINE));
         }
@@ -134,7 +134,7 @@ public class EquipmentDetailWindow : MonoBehaviour
         if (!range.IsZero())
         {
             hasPrimordial = true;
-            dps = (range.min + range.max) / 2d * weaponItem.AttackSpeed;
+            dps = (range.min + range.max) / 2f * weaponItem.AttackSpeed;
             primDps.Add(LocalizationManager.BuildElementalDamageString(dps.ToString("F2"), ElementType.VOID));
             primDamage.Add(LocalizationManager.BuildElementalDamageString(range.min + "-" + range.max, ElementType.VOID));
         }
@@ -142,7 +142,7 @@ public class EquipmentDetailWindow : MonoBehaviour
 
         if (weaponItem.PhysicalDamage.min != 0 && weaponItem.PhysicalDamage.max != 0)
         {
-            dps = (weaponItem.PhysicalDamage.min + weaponItem.PhysicalDamage.max) / 2d * weaponItem.AttackSpeed;
+            dps = (weaponItem.PhysicalDamage.min + weaponItem.PhysicalDamage.max) / 2f * weaponItem.AttackSpeed;
             //physDPS = "Physical DPS: " + dps.ToString("F2") + "\n";
             infoText.text += "Phys. DPS: " + dps.ToString("F2") + "\n";
             physDamage = "Phys. Damage: " + weaponItem.PhysicalDamage.min + "-" + weaponItem.PhysicalDamage.max + "\n";
